pyproject.toml
[project]
name = "joget-form-generator"
version = "0.1.0"
description = "Schema-driven form generation for Joget DX Enterprise Edition"
authors = [{name = "Aare Laponin"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["joget", "form", "generator", "low-code", "workflow", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "jsonschema>=4.20.0",
    "Jinja2>=3.1.6",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "PyYAML>=6.0.1",
    "mcp>=1.0.0",            # Model Context Protocol SDK
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",  # Async test support for MCP
    "black>=24.1.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "datamodel-code-generator>=0.25.0",
]

[project.scripts]
joget-form-gen = "joget_form_generator.cli.main:app"
joget-form-mcp = "joget_form_mcp.cli:main"

[project.urls]
Homepage = "https://github.com/aarelaponin/joget-form-generator"
Documentation = "https://github.com/aarelaponin/joget-form-generator/tree/main/docs"
Repository = "https://github.com/aarelaponin/joget-form-generator"
Issues = "https://github.com/aarelaponin/joget-form-generator/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
joget_form_generator = ["patterns/templates/*.j2", "schema/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/joget_form_generator --cov-report=term-missing --cov-report=html"
markers = [
    "integration: marks tests as integration tests (skip on CI)",
]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
